Move RaceViewerScreen to seperate file
